+++ src/index.ts
@@ -16,7 +16,6 @@ import * as boundedJoinSemilattice from './BoundedJoinSemilattice'
 import * as boundedLattice from './BoundedLattice'
 import * as boundedMeetSemilattice from './BoundedMeetSemilattice'
 import * as category from './Category'
-import * as choice from './Choice'
 import * as comonad from './Comonad'
 import * as compactable from './Compactable'
 import * as console from './Console'
@@ -149,10 +148,6 @@ export {
    * @since 2.0.0
    */
   category,
-  /**
-   * @since 2.0.0
-   */
-  choice,
   /**
    * @since 2.0.0
    */